Renamce check_drm.h to CheckDrm.h
:Release Notes:

:Detailed Notes:

:Testing Performed:
No tests

:QA Notes:

:Issues Addressed:
